refactor: update data structures and unit system architecture
